Share timer implementation
