+++ vnext/Desktop/React.Windows.Desktop.vcxproj
@@ -236,7 +236,7 @@
     <ClCompile Include="CxxReactWin32\JSBigString.cpp" />
     <ClCompile Include="JSBigStringResourceDll.cpp" />
     <ClCompile Include="module.g.cpp" />
-    <ClCompile Include="Modules\TimingModule.cpp" />
+    <ClCompile Include="..\Microsoft.ReactNative\Modules\TimingModule.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
@@ -304,7 +304,7 @@
     <ClInclude Include="..\Microsoft.ReactNative\RedBoxErrorFrameInfo.h" />
     <ClInclude Include="..\Microsoft.ReactNative\RedBoxErrorInfo.h" />
     <ClInclude Include="JSBigStringResourceDll.h" />
-    <ClInclude Include="Modules\TimingModule.h" />
+    <ClInclude Include="..\Microsoft.ReactNative\Modules\TimingModule.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="BeastWebSocketResource.h" />
   </ItemGroup>
