+++ vnext/Desktop/React.Windows.Desktop.vcxproj.filters
@@ -128,7 +128,7 @@
     <ClCompile Include="module.g.cpp">
       <Filter>Generated Files</Filter>
     </ClCompile>
-    <ClCompile Include="Modules\TimingModule.cpp">
+    <ClCompile Include="..\Microsoft.ReactNative\Modules\TimingModule.cpp">
       <Filter>Source Files\Modules</Filter>
     </ClCompile>
     <ClCompile Include="WebSocketResourceFactory.cpp">
@@ -184,7 +184,7 @@
     <ClInclude Include="JSBigStringResourceDll.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Modules\TimingModule.h">
+    <ClInclude Include="..\Microsoft.ReactNative\Modules\TimingModule.h">
       <Filter>Header Files\Modules</Filter>
     </ClInclude>
     <ClInclude Include="..\Microsoft.ReactNative\RedBoxErrorFrameInfo.h">
